post(phil): Introducing Purposes of Class
# Introducing purposes of types/classes

When I first learned about a Class in VB6 back in high school, they were treated as built-in primitives enabling us to write a GUI calculator. Then when I learned about User-defined Classes in Java, I was told that they are "Blueprints" for objects. Only after years of exploring different programming paradigms and languages, up and down different layers of abstractions, have I finally become somewhat confident in understanding what a Class, or in other paradigm, a Type, is. This is a sketch about how to introduce the concept of a Class or a Type better to someone without my experience, and learning Python, Java, or C++ as a beginner.

First we need to acknowledge all the built-in classes: integer (int), floating point number (float), string (str), list, array, etc. They are useful for representing values we care about. We can use integer to count the number of command line parameters. We can use float to measure the height of a building. We can use string to record names of classmates. They are also useful for combining or otherwise processing values. We can sum up the total height of a building from that of each floor. We can split a multiline string into a list of lines.

At this time usually I see the concept of a Class being introduced. We can group a useful representation, with useful procedures, into a Class. Indeed, that is how the built-in classes were made. Students would follow through the examples. But then, I have been asked often, when shall a Class be created?

For Java and C++, classes are usually created out of the necessity to group related variables. For Python, that is the job of a dictionary. Even then, on the scale of student programs, grouping related variables is not a necessity at all. To the contrary, converting the variables from individually defined to grouped is itself a point of confusion. Therefore, I think introducing Class as a grouping concept to begin with is a mistake.

A Class, or a Type, is first and foremost a differentiation. We create a new Class, because it is different from any other Class we have, somehow. The simplest example is measurements with units. A float can represent the height of a building in meters; it can also represent the height of a table in centimeters. They are different, in the sense that if we add the two floats together, without proper unit conversion, we cannot ourselves justify what the result means. At least we need to create a Class for one of meter and centimeter, so as to prevent ourselves from adding a mix of the two. Also need to note type aliasing usually doesn't work for this purpose. A whole subfield of dimensional analysis can be laid out from this example for curious and advanced students. 

From there, we can discuss different kinds of Classes, from different perspectives. For example, we can discuss combining Classes, for which the subfield of algebra can offer advanced exploration. 

Adding more member variables to a Class is taking the product of the Classes of those member variables. This is the grouping related variables aspect mentioned previously. By framing it as a combination, we can explain the necessity of using dot notation to access member variables as a separation, and hopefully more students can understand it. We can also motivate this by converting functions taking common parameters into methods of a Class with those parameters as member variables. 

Adding a super-Class from which variants of it inherit from is taking the sum of the Classes of those variants. The usual explanation of inheritance as sub-categories of a category of objects provides little guidance on what a good hierarchy of categories look like. By building the hierarchy from bottom up, hopefully we can show how super-Class arises from a need to group together different branches and put off the branching decision. This also provides an opportunity to frame `null` as the universal sub-Class, hopefully demystifying it a bit.

Another perspective may be how many possible values (cardinality) can a Class contain, for which more advanced students can go into number theory.

For integer, this is where we introduce binary numerals. For float, this where we introduce IEEE standard. For string, this is where we introduce ASCII and unicode. We can maybe even give a homework to write code to print every instance of a primitive to help students get a feel of the speed of modern computers and the limitation of it.

But then comes boolean, which contains only two values, and `null`, which contains only one value. This is where we can link the discussion back to the sum operation, and mention that since null is sub-Class of all Classes, every Class contains at least two values: itself and `null`. This is why `null` is used to represent the non-existence of a Class. 

Now we can ask what's the effect on the number of possible values. The cardinality of a super-Class would be the sum of that of its sub-Classes. The cardinality of a Class would be the product of that of its member variable Classes.

Afterwards, we can contemplate what a Class containing no possible values may look like and mean. Is it void? Does a statement return void? Toward the other direction, what about a Class which contains infinitely many possible values? From here we can introduce linked list, tree, and graph. 

Tags: compsci, edu, java