Inconsistencies in height of <body> calculation
A Kevin Powell video showed up in my YouTube feed with a fun CSS challenge from CSSBattle. I tried to follow the first challenge and got royally confused by the different result I got from display grid and place items center. Turns out Safari and Firefox do not set the height of body to the height of its containing iframe like Chromium does.

In addition, Safari and Firefox looks the same but are actually different. Chromium centers the div vertically presumably because it computes the height of body to be the same as the containing iframe. Firefox doesn't presumably because it computes the height of body to be the height of the contents of body. Safari computes the height of body to be the containing iframe, yet doesn't center the div vertically!

I know there are inconsistencies between browsers, but I did not expect inconsistency in such a simple case. Isn't this regulated by the CSS Spec?

Tag: css